+++ hwpe-tb/sw/aes_hwpe.h
@@ -3,7 +3,7 @@
 #ifndef __AES_HWPE_H__
 #define __AES_HWPE_H__
 
-void aes_hwpe_configure(unsigned int *input, unsigned int *output);
+void aes_hwpe_configure(uint8_t *input, uint8_t *output);
 
 void aes_hwpe_start(void);
